requirements.txt
astroid==3.3.3
asttokens==2.4.1
blinker==1.8.2
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
executing==2.1.0
Flask==3.0.3
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==2.1.1
opencv-contrib-python-headless==4.10.0.84
packaging==24.1
parso==0.8.4
pika==1.3.2
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
pylint==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests
six==1.16.0
stack-data==0.6.3
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
gunicorn
flask-mail
timedelta